uvcgan/data/datasets/external/PLB/regression/pyproject.toml
[project]
name = "plbregression"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Marek Bukowicki", email = "m.bukowicki@uw.edu.pl" }
]
requires-python = ">=3.11"
dependencies = [
    "albumentations[headless]>=2.0.8",
    "hydra-core>=1.3.2",
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "omegaconf>=2.3.0",
    "pandas>=3.0.2",
    "scipy>=1.17.1",
    "timm>=1.0.25",
    "torch>=2.10.0",
]

[project.scripts]
plbregression = "plbregression:main"

[build-system]
requires = ["uv_build>=0.10.8,<0.11.0"]
build-backend = "uv_build"
